Pre-defined sets of metrics

Pre-defined sets of point cloud metrics composed of different metrics_* functions available in the package.

Usage

metrics_set1(
  z,
  zmin = NA,
  threshold = c(2, 5),
  dz = 1,
  interval_count = 10,
  zintervals = c(0, 0.15, 2, 5, 10, 20, 30)
)

.metrics_set1

metrics_set2(
  z,
  ReturnNumber,
  NumberOfReturns,
  zmin = NA,
  threshold = c(2, 5),
  dz = 1,
  interval_count = 10,
  zintervals = c(0, 0.15, 2, 5, 10, 20, 30),
  KeepReturns = c(1, 2, 3, 4)
)

.metrics_set2

metrics_set3(
  x,
  y,
  z,
  i,
  ReturnNumber,
  NumberOfReturns,
  zmin = NA,
  threshold = c(2, 5),
  dz = 1,
  interval_count = 10,
  zintervals = c(0, 0.15, 2, 5, 10, 20, 30),
  pixel_size = 1,
  vox_size = 1,
  KeepReturns = c(1, 2, 3, 4)
)

.metrics_set3

Format

An object of class formula of length 2.

An object of class formula of length 2.

An object of class formula of length 2.

Arguments

zmin

Minimum height. If set, heights below are ignored in calculations.

threshold

Threshold height(s). See metrics_percabove.

dz

layer thickness to use when calculating entropy and VCI.

interval_count

Number of intervals used to divide the point height distribution. See metrics_canopydensity.

zintervals

Height intervals. See metrics_interval.

ReturnNumber

return number

NumberOfReturns

number of returns

x, y, z

X, Y, Z coordinates of the point cloud

i

Intensity

pixel_size

pixel size for calculating rumple index

vox_size

voxel size for calculating voxel metrics

Details

The three pre-defined sets of metrics include different number of the metrics_* functions and are provided to conventiently compute most of the metrics commonly used at the same time.

metrics_set1()

Included metrics: metrics_basic(), metrics_percentiles(), metrics_percabove(), metrics_dispersion(), metrics_canopydensity(), metrics_Lmoments(), metrics_lad(), metrics_interval().

metrics_set2()

Included metrics: All metrics in metrics_set1(), metrics_echo(), and metrics_echo2().

metrics_set3()

Included metrics: All metrics in metrics_set2(), metrics_kde(), metrics_HOME(), metrics_rumple(), metrics_voxels()

Currently no set includes metrics_texture().